+++ src/main.js
@@ -10,11 +10,15 @@ import { fas } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/vue-fontawesome"
 
 import VMdEditor from '@kangc/v-md-editor'
+import VMdPreview from '@kangc/v-md-editor/lib/preview'
 import '@kangc/v-md-editor/lib/style/base-editor.css'
+import '@kangc/v-md-editor/lib/style/preview.css'
+import enUS from '@kangc/v-md-editor/lib/lang/en-US'
 import vuepressTheme from '@kangc/v-md-editor/lib/theme/vuepress.js'
 import '@kangc/v-md-editor/lib/theme/style/vuepress.css'
 import Prism from 'prismjs'
 import 'prismjs/components/prism-json'
+import mdplugin from './mdplugin'
 
 import App from './App.vue'
 import router from './router'
@@ -23,8 +27,18 @@ import { QwikiPlugin } from './plugin'
 
 library.add(fas)
 
+VMdEditor.lang.use('en-US', enUS)
 VMdEditor.use(vuepressTheme, {
-  Prism,
+    Prism,
+    extend(md) {
+        md.use(mdplugin)
+    },
+})
+VMdPreview.use(vuepressTheme, {
+    Prism,
+    extend(md) {
+        md.use(mdplugin)
+    },
 })
 
 const app = createApp(App)
@@ -40,5 +54,6 @@ app.use(Oruga, {
 app.use(QordialPlugin)
 app.use(QwikiPlugin)
 app.use(VMdEditor)
+app.use(VMdPreview)
 
 app.mount('#app')